requirements.txt
chardet==5.1.0
deap==1.3.3
ipython==8.12.0
matplotlib==3.7.2
natsort==8.4.0
nltk==3.8.1
numpy==1.24.4
pandas==2.0.3
regex==2022.3.2
scikit_learn==1.2.2
scipy
tqdm==4.65.0
pytorch_pretrained_bert==0.6.2
